Fix ShadowGradientDrawable
This shadow was causing test failures, because the getColor() method has the same name as an Android SDK method that was added with api version 24 but a different return type. The shadows method returns int whereas the Android SDK method returns ColorStateList. If the code under test is using the Android SDK method the test will fail, because it cannot find a method with this name and return type on the Shadow.

This change fixes the issue by renaming the affected method on the shadow.

PiperOrigin-RevId: 217111981