Add shadow for androidx.multidex
This is a workaround for the broken Robolectric support in the support library. At one point Robolectric was supported but internal changes to Robolectric broke that support.

I (jongerrish@) propose fixing this in https://issuetracker.google.com/issues/117224552 at which point custom shadows of this support library should not be necessary

PiperOrigin-RevId: 217936979