Support dealing with v53 (Java 9) .class files.
This is a backport of a CL already submitted internally. More specifically:
 * I submitted the original CL internally (internal CL 215005271)
 * which was pushed to github:
   https://github.com/robolectric/robolectric/commit/8334ae81805ce7d13898f6438ce0fee98fef1c0f
 * which I then cherry-picked here.

This commit fixes https://github.com/robolectric/robolectric/issues/3999 but regresses https://github.com/robolectric/robolectric/issues/2442

Original commit message follows:

The Java VM Specification says that updating final fields from the putstatic bytecode is only allowed in <init> or <clinit> of the current class [1]. OpenJDK 9 enforces this for code from v53 (OpenJDK 9) .class files [2].

Since Robolectric updates various fields from $$robo$init or from __staticInitializer__, those fields must be nonfinal in order for Robolectric to work on v53 fields.

Before this CL, Robolectric supported running under OpenJDK 9 java, but not on v53 .class files (written by default by OpenJDK 9 javac). This CL fixes one breakage on v53 .class files, although other breakages might remain in areas that I have not exercised; for example, there are still a couple of occurrences of Opcodes.ASM4 in sources and of org.ow2.asm:asm-tree:5.0.1 in buildSrc/build.gradle, which might need to be updated to ASM6, but I don't know if / where they are used.

Note that this CL regresses on the bugfix for https://github.com/robolectric/robolectric/issues/2442 because that fix relied on __robo_data__ being final (which is no longer supported). The original bug report involved @InjectMocks on an Activity, which is probably an edge case that we may not want to support. If we do want to support it, then we'll have to find a different fix for the original issue; for example, if there was a setter for __robo_data__ whose implementation does nothing then mockito would probably call that instead.

[1] https://bugs.openjdk.java.net/browse/JDK-8157181
[2] http://hg.openjdk.java.net/jdk9/jdk9/hotspot/annotate/c558d46c1af2/src/share/vm/interpreter/linkResolver.cpp#l960

Bug: 116872534
Test: Checked that the following still passes after this CL, both in
      workspaces with and without the environment variable
      EXPERIMENTAL_USE_OPENJDK9 set to true:

   make RunCarSetupWizardRoboTests RunCarSetupWizardRoboTests-jacoco \
     RunCarSetupWizardLibRoboTests RunCarSetupWizardLibRoboTests-jacoco
     RunCarSettingsRoboTests RunCarSettingsRoboTests-jacoco \
     RunCarMessengerRoboTests RunCarMessengerRoboTests-jacoco \
     RunSettingsGoogleRoboTests

Change-Id: I333a23b9b58689d939222650df5dbd162108dbc0