Fix NPE when laying out views
A recent change added mPendingInsets to ViewRootImpl. It is null after
ViewRootImpl's construction and remains so until layout is requested.

Since layout is normally a posted operation in true Android the nullity
is not an issue. In robolectric, since the posted layout pass occurs
immediately, an attempt to use a null mPendingInsets reliably results in
the NPE.

Change-Id: I47d4e6748c7cbb2422d31977de794fc311323279
Fixes: 120245685
Test: make -j56 RunCarSetupWizardRoboTests