Special-case ShadowParcel's handling of zero bytes.
In almost all cases, we want ShadowParcel to behave strongly-typed, to catch
bugs in production code and tests and to detect protability issues.  However,
Parcel's handling of values that are entirely zero is well-defined, because all
zeroes are interpreted the same whether big-Endian or little-Endian and have
special meanings.  There are some reasonable use cases:

- Tests for a parceled object with a marshalled byte stream of all zeroes.
This returns default values for primitives and empty for strings/arrays, and is
a popular way to test BadParcelableException.

- Writing a long and reading two ints, or writing two ints and reading a long.

- Writing a byte array of all zeroes, and reading the zeroes (not the length)
as other data types.

This only supports values that are entirely zero from begin to finish, and does
not support reading strings as other zero-able types.  This is partially to
simplify the implementation, and also because it limits the zeroes behavior to
cases where it is more likely to be intentional rather than a bug in a test.

The implementation works by:

- At write time, flagging values that are represented as zeroes.

- At read time, before consuming certain types, check if there are enough
consecutive zeroes to parse the desired type as zero.

PiperOrigin-RevId: 227907719