Copybara import of the project:
--
227943230 by romkal <romkal@google.com>:

Migrate addResolveInfoForIntent into the new approach with filters.

This change makes addResolveInfoForIntent to behave in a consistent way with the rest of package manager functionality.

If you add a resolveInfo for intent you can query the component that gets resolved, you can modify the intent in a way that should still resolve the component, you can query the package for the resolved component and its applicationInfo will be consistent with whatever is in the package.

In order for it to be reasonable backward compatible, some assumptions had to be made:
1. When component doesn't have name or package, some defaults are added assuming, that every newly added component is in different package.
2. If existing package is used its existing applicationInfo will be used instead of the one in the resolveInfo
3. If no component is specified at all it is assumed it is an activity.
4. If existing component is used (name match), its original ComponentInfo will get used instead of the one in the resolveInfo.
5. It is assumed, that the component has a intent filter that matches as strictly as possible the intent provided.
6. Component name/package specified in the intent has preference over what is specified in the component

In summary components already in the system have preference over what is specified in the resolveInfo.
Intent data has preference over what gets specified in the resolveInfo.

PiperOrigin-RevId: 227943230