Move intent filters to their own lists and out of the Package. [Reapply after rollback]
This is done to implement more natural intent resolution mocking approach, where instead of doing
addResolveInfoForIntent(Intent, ResolveInfo) one can do addIntentFilterForComponent(ComponentName, IntentFilter).

It makes clear what the filter is, what is the type of the component (as it has to be added first), sets correct filter field in ResolveInfo and is resilient to changes in the intent (all intents that should match, will match - not only the one registered).

This has additional effect of correctly returning component metadata from queryIntentActivities().

Ultimately everything should work the same way regardless of the source of the component (manifest file or test code).

This is another attempt to submit this after series of rollbacks.

PiperOrigin-RevId: 229496278