Set AppBindData#processName consistently with ApplicationInfo#processName.
Instead of using the hard-coded "org.robolectric" value, we should use the package name of the app. This makes Application#getProcessName, a new API added in Android P, return a consistent value.

This logic still doesn't quite match the platform's - in particular, it ignores the process attribute that can be set at the <application> level of an AndroidManifest.xml file - but it works for most apps.

PiperOrigin-RevId: 230566222