Fix Bug in GenerateReceiver/ReceivesAccount
Proposed fix for b/124361003

Instead of calling goAsync() in handleUnorderedBroadcast() call it from onReceive().

The existing code only goes async with the unordered variance so in order to preserve this behavior, i had to be able to tell if the receiver is ordered or not.

In the original code this is done by calling orderedReceiveForGeneratedCodeOnly() and if it returns null, it means it's not ordered.

But orderedReceiveForGeneratedCodeOnly() requires an entry point as an an arg and we don't have one yet in onReceive().

We could call it with null and catch the exception but that would not be very nice. Rather, I changed the generator to add a isOrderedReceiver(). Is there a better way to tell that I missed?

The Copybara line is so I can run g4 presubmit -a mail. Needed because of the diffbase CL.

PiperOrigin-RevId: 233730604