Unset all session timers for a behavior handler
Yes there is a tri-diffbase to be able to test this :P It goes:
Base CL: Shadow for session timers
Diffbase on Base CL: Session timer tests that were approved Friday, but can't be submitted until the Shadow is submitted

This is to make it easier to see just the new tests added for this method to ensure coverage.

I plan to remove the tests and diffbase from this CL, and submit this once it's been reviewed. I will then put all tests in one XL CL to submit at whatever point the Shadow is in.

PiperOrigin-RevId: 235911275