Introduce a new ?Realistic looper? threading model.
Realistic looper is a new set of simplified APIs and shadows for controlling tasks posted to Android loopers. Realistic looper is opt-in for now, but we hope to make it the default in an upcoming release and deprecate the existing Scheduler/ShadowLooper APIs.

The major differences compared with the existing threading model are:
 - Tasks run on the correct thread. Just like in real Android, each Looper has its own thread
 - The current ShadowLooper/Scheduler supports multiple execution modes: PAUSED, UNPAUSED, etc. In Realistic looper (for now), the main looper is effectively always paused. Non-main loopers behave like they would on a real device, with posted tasks executed asynchronously.
 - There is a single system time, controlled by ShadowRealisticSystemClock. Previously each scheduler had its own time, which could be incremented independently

To use:
 - Apply the @LooperMode(PAUSED) annotation to your test class / method
 - Replace existing Scheduler or ShadowLooper calls with shadowBaseLooper().* calls. ShadowBaseLooper will automatically call either
 the legacy ShadowLooper APIs or ShadowRealisticLooper APIs based on the system property.

You should see a custom error message when tests fail with unexecuted runnables in the main looper, to highlight that as a potential cause.

Legacy Scheduler etc APIs will throw when called when looper mode != LEGACY

See #3369 for additional background.

PiperOrigin-RevId: 241162790