Fix activities leaking between test cases.
This fix will stop leaving an activity as strongly reachable once the shadows are reset.

Note that, as of now, calling ActivityController#destroy() in a single test still leaves that activity strongly reachable until the shadows are reset.
There seems to be some cleanup around the window performed in ActivityThread#handleDestroyActivity() that we would need to emulate here for that to happen.

PiperOrigin-RevId: 245463548