Fall back to Android's implementation of PackageManager.getPackageArchiveInfo
This method is intended to parse apks on disk, and doesn't require that the apk
be installed. However, robolectric's implementation only works for apks that
are associated with an installed app.

This change maintains the current logic for returning the "installed"
PackageInfo for installed apks, but falls back to the native implementation
for non-installed apks, which should parse the apk and generate a PackageInfo as
appropriate.

This change only affects the behavior in binary resource mode. In legacy
resource mode, this method will still return null for a not-installed apk.

PiperOrigin-RevId: 245836995