Modifies ShadowApplication and ShadowInstrumentation so that if configured, unbindService does not call onServiceDisconnected.
The fact that Robolectric improperly calls onServiceDisconnected makes it impossible to test code that calls unbindService after an onServiceDisconnected callback.

This change modifies the appropriate classes to allow users to disable this incorrect behavior with the expectation that in the future, the default behavior will switch to the correct behavior, and then ultimately the option to enable the incorrect behavior will go away.

Fixes https://github.com/robolectric/robolectric/issues/3882

PiperOrigin-RevId: 248366218