Implement getActiveSubscriptionInfoCountMax().
It looks like the real implementation binds to a service, so it's not clear
exactly how this function works; I've based my implementation on the javadoc.

Because it looks like it's possible to set the info list to null in the shadow,
I've accommodated this behavior when evaluating list size.

It doesn't look like it's ever possible in the shadow for there to be a difference
between Max() parameters, but I've kept it to maintain consistency with the javadoc.

PiperOrigin-RevId: 249110037