Add a supplier-version of registerInputStream/registerOutputStream
ShadowContentResolver behaves differently than the system ContentResolver in one important regard: Once the client closes the stream, that stream will then be closed in all subsequent calls. Example:

```
    shadowContentResolver.registerInputStream(
        uri21, new ByteArrayInputStream("ourStream".getBytes(UTF_8)));
    InputStream inputStream1 = contentResolver.openInputStream(uri21);
    byte[] data1 = new byte[9];
    inputStream1.read(data1);
    inputStream1.close();
    InputStream inputStream2 = contentResolver.openInputStream(uri21);
    byte[] data2 = new byte[9];
    inputStream2.read(data2); // <--- this will throw exception
```

The solution is to provide an alternative overload for registerInputStream/registerOutputStream, which takes a Supplier, a lambda which returns a stream. In this way, clients which require the ability to have openInputStream() return an opened input stream multiple times for the same uri, can provide a Supplier which returns a new open instance.

PiperOrigin-RevId: 255422684