Adding ShadowNotificationManagerCompat with the ability to test areNotificationsEnabled()
The implementation of areNotificationsEnabled() will do the following for the given SDK version:
24+: send the call to the NotificationManager, which can be picked up by ShadowNotificationManager.
19-23: return the value set in the ShadowNotificationManagerCompat - this is where this shadow is working.
18-: return true, same as the real object.

This also allows one to set which ShadowNotificationManagerCompat is returned when NotificationManagerCompat.from(Context) is called. This allows one to force the use of the set up Shadow.
Calling NotificationManagerCompat.from(Context) does not always return the same instance if this is not set.

PiperOrigin-RevId: 263227904