Fix triggering broadcasts from explicit intents if data on intent is specified.
Until now, a broadcast in a manifest would only be triggered by an explicit intent if the data of the intent was empty. According to https://developer.android.com/guide/components/intents-filters#Types an explicit intent is always triggered as long as the component name matches, even if there is no intent filter.

PiperOrigin-RevId: 272627681