Add shadow implementation of MediaBrowserCompat#getSessionToken()
Shadowed version of MediaBrowserCompat isn't relying on the actual token, rather it maintains a list of media items itself. However, as it doesn't use the actual session, getSessionToken() doesn't return meaningful value, and even worse it will cause crash if the code being tested is using this API. This CL makes it return a fake media session token though it won't be used.

PiperOrigin-RevId: 275069739