Do not check for queued runnables when test is aborted with AssumptionViolatedException
When a JUnit test throws an AssumptionViolatedException, the test runner should not generate a test case failure: see https://junit.org/junit4/javadoc/latest/org/junit/AssumptionViolatedException.html.

However, when a Robolectric test running with an AndroidTestEnvironment throws an exception (including an AssumptionViolatedException), the AndroidTestEnvironment#checkStateAfterTestFailure method checks if there are queued runnables and itself throws an Exception if there are. The idea is to provide a hint to the developer that the queued runnables may be the cause of the test failure.

If a test is aborted with an AssumptionViolatedException, though, Robolectric should allow the test to be aborted without introducing its test case failure. This change does that by not calling AndroidTestEnvironment#checkStateAfterTestFailure where a test throws an AssumptionViolatedException.

PiperOrigin-RevId: 281343466