Add the ability to set Layout for ShadowTextView
Our team has been working on a custom implementation of EditText which
includes a custom MovementMethod. In some cases, we default to the
default implementation of movement, which ends up calling up to
Selection.moveLeft, Selection.moveRight, etc. Similarly some
calculations require the Layout. We'd like to be able to set a mocked
Layout instance for this.

This change includes the needed changes to ShadowTextView and a test to
verify functionality.