Modify canNotifyAsPackage method in ShadowNotificationManager to support tests.
Modifies canNotifyAsPackage method implementation in ShadowNotificationManager to support unit testing. This does not provide a complete implementation of canNotifyAsUser because the ShadowNotificationManager doesn't know how to work with contexts associated with other packages (via createPackageContext). So we are providing a new method is ShadowNotificationManager (setNotificationDelegateForPackage()) which will keep track of the mapping between app packages and its delegates.

PiperOrigin-RevId: 305410273