Allow updating history in Robolectric ShadowWebView async.
WebView does not update its history inline when loadUrl() is called. It starts a "navigation" which might or might not "commit" (i.e. decide to show something to the user). For example, a gen204 request never commits, calling loadUrl() with such a request would never "commit" to the history. Or two loadUrl() calls can be made immediately after one another, in which case the first is never committed as it is dropped. This commit always happens async and never inline. As such, ShadowWebView behaviour is inconsistent with the prod behaviour.

Introducing pushEntryToHistory() API allows this commit to be performed asynchronously when requested. This API also takes a URL, which does not have to be the same as the one in loadUrl() as there might have been e.g. server-side redirects until the page is committed. Though if the URL does not go through any changes the caller can also perform pushEntryToHistory(webView.getUrl()).

This API can also be used without an associated loadUrl() call to simulate a renderer-initiated navigation, such as when the user-clicks on a link.

PiperOrigin-RevId: 311727608