Add observer APIs for Activities and phone calls
Add interface ShadowInstrumentation.StartActivityListener to enable tests to register observers that get invoked when the code under test requests to start an Activity.

Add interface ShadowTelecomManager.CallListener to enable tests to register observers that get invoked when the code under tests requests to place an outgoing phone call or when an incoming call is added.

PiperOrigin-RevId: 312328332