Remove pegdown-doclet markdown javadoc doclet plugin
Robolectric was using the pegdown-doclet gradle plugin to support Markdown in
Javadoc. This plugin is not compatible with JDK 10+, which introduces a new
doclet API and javadoc command-line arguments. I tried to find alternatives to
pegdown-doclet, but couldn't find any that support JDK 10+. Instead, remove the
plugin and switch to using vanilla Javadoc. This may require some follow-up
commits to reformat markdown javadoc into HTML.

Fixes #5743