Drop maven-ant-tasks dependency and fetch from Maven manually
The dependency maven-ant-tasks is outdated and has security issues. Replace the
maven dependency fetching logic with simple fetching through HTTPUrlConnection.
There was an initial attempt to use maven-resolver, but the version of Apache
HttpClient that uses conflicts with the HttpClient that must be bundled for
Android 16-18. Because the use cases for Robolectric is very narrow (only
fetching single artifacts with no dependencies), it is feasible to use this
approach.