Use loose signatures on ShadowTelephonyManager.requestCellInfoUpdate
This is to avoid referring to Android Q's CellInfoCallback in the method
signature. The problem was related to the inner anonymous runnable classes
referring to the outer CellInfoCallback. Instead, perform the casting inside
each anonymous Runnable.

For some reason this does not seem to be an issue on some JVMs.

PiperOrigin-RevId: 325539220