Update ReflectionHelpers to only remove final modifier for final fields
Previously, calling Reflecion.getStaticField or
ReflectionHelpers.setStaticField would always remove the 'final' modifier using
reflection, even for non-final fields. This causes a crash in JDK13. Instead,
only remove the 'final' modifier for fields that are actually declared final.
This avoids the illegal reflective operation in Robolectric itself.

In an upcoming release we can look for alternatives to Reflection for getting
rid of final modifiers, such as Unsafe.

This is related to #4776.

PiperOrigin-RevId: 324612122