Use Sun JSSE provider for Maven TLS artifact fetching
If a Maven artifact fetch was performed after the BouncyCastle JCE security
provider was installed, a cryptic error message could occur:

    java.io.IOException: stream does not represent a PKCS12 key store

The problem was that BouncyCastle JCE changes the default KeyStore format from
JKS to PKCS12. If the default SunJSSE was used to establish a TLS connection,
it expected the KeyStore to be in JKS format, which caused the exception.

Instead, switch to using the built-in Sun JSSE provider, which works with a JKS
key store..

Longer term we should re-evaluate which providers should be installed and
perhaps use a custom Robolectric provider.