Fix Robolectric ShadowProcess#setThreadPriority's priority param range
The current acceptable range is of values is [-19, 19]
It should be [-20, 19] as documented by android.os.Process#setThreadPriority(int, int).

Here the usage of Process.THEAD_PRIORITY_* constants is confusing.
Process.THREAD_PRIORITY_LOWEST = 19
Process.THREAD_PRIORITY_URGENT_AUDIO = -19
Firstly, Process.THREAD_PRIORITY_LOWEST is the highest number accepted, which is misleading without additional context.
Secondly, the range of acceptable priority values is not tied to these constants. In other words, "magic numbers" i.e. -20 and 19 are more useful in this scenario as they tell us more about the priority numbers.

For these reasons, the raw integer values are used since that's what we really need to check for--that the values should be in the range [-20,19]. Not between the highest and lowest Process.THREAD_PRIORITY_* constants.

PiperOrigin-RevId: 327707093