Clear Activity DecorViews from WindowManager when they are destroyed
Previously, when Activities were destroyed in Robolectric, their state in
WindowManager was not cleared. This caused a leak during the course of a test
run. While the state did eventually get cleared after the test was over, it
caused problems with tests that allocated a large number of activities within
the same test.

With this change, when an activity is destroyed, View.onDetachedFromWindow
will be triggered on the view hierarchy. This may cause issues for tests that
are not expecting View.onDetachedFromWindow to get called. However, this is
only likely to affect a tiny fraction of tests.

Fixes #5530

PiperOrigin-RevId: 341913086