Cancel ShadowView AnimationRunner when View.clearAnimation is called
Previously, View.clearAnimation would not cancel the animation that was set in
View.setAnimation. This caused animations specified in View.setAnimation to be
run to completion regardless of whether they were cleared or not.

This could be a performance problem if tests used a combination of
View.setAnimation and ShadowLooper.idle.

PiperOrigin-RevId: 342091936