Add ShadowMagnificationController for N+
A single instance of MagnificationController exists per AccessibilityService.
It abstracts over sending commands to the framework over Binder. These commands
allow the client to magnify areas of the screen, and to query if the screen is
currently being magnified.

Before this shadow existed, using MagnificationController from Robolectric
would have no visible effects. i.e., if you were to call .setScale(n, animate)
to modify the magnification scale, then calling .getScale() would not return n:
it would instead return the default scale (1f). This shadow keeps internal
magnification state, so that the APIs behave as expected.

Currently an assumption baked into this code is that only one
AccessibilityService is active at once. This should be a safe assumption in
general, as there are various disadvantages to having multiple services in an
apk (from both the performance perspective, but also from user convenience, as
each service has to be activated separately, from a deeply buried Settings page).

PiperOrigin-RevId: 350651958