Fix how the OnInitListener is invoked in ShadowTextToSpeech
Previously, in ShadowTextToSpeech, the OnInitListener callback was always
called inline with the ERROR code. Instead, default to calling it
asynchronously with the SUCCESS code, which is what happens in real Android.

In the case that tests want to check the error condition, add a new API
ShadowTextToSpeech.setOnInitStatus(int) which can be used to set the status
code used in the callback.

PiperOrigin-RevId: 350776217