Avoid ClassNode construction for classes that do not get instrumented
According to https://asm.ow2.io/performance.html, using a ClassReader is a lot
faster for reading basic class metadata than creating a ClassNode. This is
because to create a ClassNode, every single byte of the class file needs to be
visited, including method instructions. For ClassReader, a lot of the work is
done lazily. Previously, a ClassNode object was constructed for every class
loaded by SandboxClassLoaer, regardless of whether or not it was instrumented.

Instead, use a ClassReader approach to determine whether or not a class needs
to be instrumented, and then build the ClassNode if it needs to be
instrumented. This should improve performance for larger tests that load a lot
of non-Android classes.

PiperOrigin-RevId: 351251828