Tweak ShadowTextToSpeech OnInitListener API
Currently, ShadowTextToSpeech's OnInitListener API is inconsistent and
confusing because the success case requires flushing the main Looper and the
error case is executed inline. This does not allow for testing some scenarios,
such as calling the error case asynchronously.

Instead, switch to using an explicit callback API. This gives tests more fine
grained control of when both the success and error modes are called.

PiperOrigin-RevId: 351512163