Remove publish_test_results job
This job is redundant, the results already get posted to the PR itself,
and the final aggregate job was only needed before individual checks
could be reported. Also the email notification is noisy. We may need an
aggregate job in the future, however, to upload snapshots for tests on
the master branch.