Read actual images in ShadowBitmapFactory
Add a method to ImageUtil that reads the whole bitmap using ImageIO (similar to the existing method that reads the image dimensions) and use it in ShadowBitmapFactory.

Fix ShadowBitmap.extractAlpha to extract ARGB values with the same alpha as the source instead of using the alpha value itself as the pixel value. This is necessary because the ctesque BitmapTest was previously getting an image with only zero pixel values from the BitmapFactory, masking the incorrect implementation. Also update ShadowBitmapTest, which was actually verifying the incorrect behavior (the fact that the new behavior is correct is verified by the ctesque test that is run on an emulator as well as under Robolectric).

PiperOrigin-RevId: 356733110