Keep alpha value to alpha channel bit for Bitmap.extractAlpha
The behavior of Bitmap.extractAlpha on the  official emulator is to use
Bitmap.Config.ALPHA_8 to store the new bitmap value with only alpha
channel. But it alpha value is on the alpha channel bit, instead of the
alpha becomes the whole pixel value. It can be tested with the following
test on the official emulator:

@RunWith(AndroidJUnit4.class)
public class ExampleInstrumentedTest {
    @Test
    public void testBitmapExtractAlpha() {
        // Context of the app under test.
        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
        Bitmap bitmap =
                BitmapFactory.decodeResource(appContext.getResources(), R.drawable.test0011ff);
        Bitmap newBitmap = bitmap.extractAlpha();
        assertEquals(Color.alpha(bitmap.getPixel(1, 1)), Color.alpha(newBitmap.getPixel(1, 1)));
        assertEquals(0xFF000000, newBitmap.getPixel(1, 1));
        assertEquals(Bitmap.Config.ALPHA_8, newBitmap.getConfig());
    }
}

So this CL restore the BitmapTest.testExtractAlpha, and fix the
ShadowBitap.extractAlpha to keep alpha value to alpha channel bit and
erase other channels' value.

Signed-off-by: utzcoz <utzcoz@outlook.com>