Improve the fidelity Canvas.drawBitmap(Bitmap, float, float, Paint)
Previously, calling Canvas.drawBitmap(Bitmap, float, float, Paint) was
effectively a no-op. It just added some metadata to the ShadowCanvas
description.

Add a basic implementation for it that attempts to copy data from the source
ShadowBitmap colors array to the destination ShadowBitmap colors array.

There also a fast path in the case that the two bitmaps have the same widths.

Also add to BitmapTest to ensure the results are consistent across real
devices.

PiperOrigin-RevId: 360230571