Clear ShadowBitmap description when eraseColor is called
There are some libraries (e.g. Glide) that maintain a pool of Bitmap objects in
order to perform transformations. Due to the reuse of Bitmap objects, the
corresponding ShadowBitmap descriptions can grow exponentially.

When these Bitmaps are returned to the pool, eraseColor(TRANSPARENT) is called
on them. Handle this case and clear the description if it is called.

PiperOrigin-RevId: 360831058