Improve fidelity of ImageView.draw(Canvas)
Previously it did not call Drawable.draw(Canvas) on the underlying Drawable
from the ImageView.

The logic to do this had to go into ShadowView as ImageView does not override
draw(Canvas). Another option was to create a ShadowImageView and extend
draw(Canvas), but that would cause problems for tests that have custom shadows
for View, which is one of the more commonly occurring shadows.

This is related to https://github.com/robolectric/robolectric/pull/6147.

PiperOrigin-RevId: 361385295