Update ShadowBitmap to be backed by Java's BufferedImage
Previously ShadowBitmap contained a int[] colors array that represented the
pixels of the image. This was low fidelity and was impossible to perform
operations like scaling. Update ShadowBitmap to be backed by Java's
BufferedImage, which is an industry standard Bitmap library.

On nice thing about this approach is that the BufferedImage that is used in
ShadowBitmapFactory to get image metadata can now be reused as the image buffer
for the ShadowBitmap. This saves some additional copying.

Some operations (such as Canvas.draw(Bitmap), and Canvas.draw(Paint)) have not
been migrated to Graphics2d yet due to slight behavioral differences with real
Android. Specifically, Graphics2d tends to simplify pixel values in ARGB_8888
For instance, values like 0x10 with zero alpha transparency get converted to 0.
This simplification does not occur in real Android.

Internally images are all stored using TYPE_INT_ARGB (ARGB_8888) format for
fast comparison (via Bitmap.sameAs) to other bitmaps.

Add a bunch of ctesque BitmapTests to capture various issues that came up to
ensure that the behavior is consistent with real Android.

PiperOrigin-RevId: 362750703