Add option to require valid image data in BitmapFactory.decode methods
It is currently impossible to write tests that depend on the error case of
BitmapFactory.decode methods. When invalid data is supplied to
BitmapFactory.decode methods in real Android, BitmapFactory is supposed to
return null and any provided BitmapFactory.Options.out{Width,Height} should be
set to -1. Currently Robolectric returns a Bitmap object with width and height
set to 100 in this case.

Add an new public API method ShadowBitmapFactory.setAllowInvalidImageData that
can be used to tweak this behavior. If
BitmapFactory.setAllowInvalidImageData(false) is called, BitmapFactory behaves
consistently with real Android by returning null and setting any provided
BitmapFactory.Options.out{Width,Height} to -1.

Long-term the goal is to BitmapFactory default is to to have
allowInvalidImageData=false, which is consistent with real Android.
PiperOrigin-RevId: 363545671