Add ContentProviderBuilder and MatrixCursorBuilder
Many of the methods in ShadowContentResolver have been deprecated in order to
discourage shallow testing of content provider queries. Add a helper class
that lets you quickly construct content provider objects that can be used in
tests.

Also add MatrixCursorBuilder, which makes it easier to construct a
MatrixCursor object for all API levels.

PiperOrigin-RevId: 364403078