Consolidate some Bitmap.createBitmap @Implementation methods
Previously there were three separate @Implementations of
Bitmap.createBitmap(Bitmap source, ...) methods.

In the Android code the 1-arg Bitmap.createBitmap(Bitmap) and 5-arg
Bitmap.createBitmap(Bitmap, int, int, int, int) just end up calling the 7-arg
variant. Remove two of the redundant @Implementation methods which will
naturally defer to the 7-arg one.

Also remove the ambiguous and unused logic that adds "created from Bitmap
object" to the record when Bitmap.createBitmap(Bitmap) is called. This
ambiguous because it changes the description of the existing source bitmap,
which leads to an incorrect description of the source.

PiperOrigin-RevId: 366525524