Lazily create the WebViewFactoryProvider proxy in ShadowWebView
Previously this was happening in ShadowWebView's clinit, which caused
NoClassDefFoundError if a custom shadow's clinit was called before the Sandbox
was initialized. This is because WebViewFactoryProvider is an internal API and
is not defined in the public SDK stubs.

It also avoids unnecessarily the proxy in the first invocation of the
ShadowWebView resetter if ShadowWebView is never used.

Fixes #6329

PiperOrigin-RevId: 368074363