Update ShadowPhone to use the call list in Phone
Instead of overriding the call list in the shadow, use the internal one
in Phone. Adds methods to the ShadowInCallService to expose handler
methods so that we may bind an InCallAdapter to the ShadowInCallService,
which allows a more realistic way of adding and updating calls using a
ParcelableCall as input.

PiperOrigin-RevId: 368480951