Prefer the stored Application and Instrumentation in ShadowActivityThread
Previously, when calling ActivityThread.getApplication or
ActivityThread.getInstrumentation, it always returned the objects from the
global ActivityThread stored in RuntimeEnvironment.getActivityThread().
Instead, prefer the ones that are stored in the associated ActivityThread
object.

This prevents potential race conditions when lazy loading is on and a
background thread calls either ActivityThread.getInstrumentation or
ActivityThread.getApplication when the global ActivityThread has not been
initialized yet.

PiperOrigin-RevId: 370299301