Use the real implementation of BluetoothAdapter.getDefaultAdapter
The previous shadow implementation of BluetoothAdapter.getDefaultAdapter
instantiated a BluetoothAdapter using Shadow.newInstanceOf, which meant none of
the important fields were being set on BluetoothAdapter. This likely resulted in
a lot of unnecessary shadow methods for BluetoothAdapter (as well as NPEs).

To make this work, two bluetooth entries had to be added to
ShadowServiceManager.

Also, deprecate ShadowApplication.getBluetoothAdapter. There's no reason to
have this API on ShadowApplication.

PiperOrigin-RevId: 371776282