Update Bootstrap display configs in setQualifiers if the application isn't already loaded
In the LazyLoad.OFF case, the Resource objects returned by
Resources.getSystem() and application.getResources() are both updated
inside RuntimeEnvironment.setQualifiers. But in the LazyLoad.ON case,
before the application is loaded, only Resources.getSystem() would get
updated. So when we lazily load the application, we also have to
propagate any changes that were made to the object inside setQualifiers
via updateConfiguration(...)

PiperOrigin-RevId: 374537697