Adding support for @ReflectorObject.
This is part of of larger initiative to replace directlyOn with reflector. The initialization is handled by ShadowWrangler and the mechanism is similar to @RealObject. The only difference is that the annotated variable will store reflector(ClassReflector.java, realObject) instead of just realObject. This avoids the overhead of creating a new object each time reflector(...) is invoked.

From microbenchmarks in ReflectorTest, caching the reflector object is around
2x faster than creating a new reflector object on each invocation.

The recommended use for this annotation is applying it to some instance variable objectReflector. Method calls by directlyOn(...) and reflector(...) can then simply be replaced by objectReflector.method(...).

PiperOrigin-RevId: 378161587