Fix memory leaks that can occur with lazy loading enabled
In particular, FontsContract and RuntimeEnvironment.systemContext were still holding onto the old application. In the eager loading case, those two references were previously be automatically updated by the test runner. For lazy loading, it has to be explicitly cleared so that it isn't inadvertently used before the new application is loaded.

PiperOrigin-RevId: 378924509