Close open CameraDeviceImpl objects after each test
Previously, when CameraDeviceImpl objects were GC'd, their close callbacks
would be executed on the finalizer thread. This would either post a task to a
Handler or Executor from a previous test and potentially cause unrelated
subsequent tests to fail.

Add a ShadowCameraManager resetter that explicitly closes any open
CameraDeviceImpl objects. This will cause any potential issues to be surfaced
immediately, especially when the executor of the close callback is a threaded
executor.

This was the root cause of the ShadowPausedLooperTest failures on GitHub CI.
These likely leaked during ShadowCameraManagerTest.

PiperOrigin-RevId: 381127583