Add integration_tests:compat-target29
This module is used to test integration compatibility problem for
project with target 29. When we add API for new SDK with new added class
on method signature, the app with target SDK less than new SDK will
encounter parsing problem, for example NoClassDefFoundError, when
integrating new robolectric version. For example, we add implementation
for LocationManager#registerAntennaInfoListener on
ShadowLocationManager, and this method has one input parameter with type
GnssAntennaInfo.Listener, added from API 30. If the app with target 29
try to use Robolectric to test LocationManager, the runtime will throw
NoClassDefFoundError for tests. This integration test is used to test
compatibility problem when adding those methods to avoid those type
problems occur on user's probject.

Signed-off-by: utzcoz <utzcoz@outlook.com>