Improve reflector caching using a strong Map
Previously, cached reflector classes were always eligible for GC because they
were stored in a weak-value Map. At P95 tests were spending up to 500ms
redefining reflector classes. Update to use a concurrent HashMap instead of a
weak-value HashMap. There are ~130 reflector classes, all of which are created
lazily, so it is not expected to contribute much to memory usage.

PiperOrigin-RevId: 385674538