Fix SQLiteDatabaseTest foreign key test to match Android behavior
SQLiteDatabaseTest.shouldThrowWhenForeignKeysConstraintIsViolated failed if run
on a real Android. It expected that a foreign key exception be thrown. In real
Android, however, no exception was thrown. Only a message is printed in logcat.

Update the test to delete a row that has a reference to another table. This
causes an exception in both real Android and Robolectric. Move this to a
ctesque test to ensure fidelity with real Android.

PiperOrigin-RevId: 386770883