Add parameterized test name method to the test runner
Previously this was just using the default implementation of the test name method which returns the method name, this means each version of the parameterized test has the same test name. This eventually gets populated into the description passed into e.g. the rules.

This change makes this runner consistent with junit Parameterized runner.

PiperOrigin-RevId: 392678724