Update default SQLite synchronous mode when WAL is enabled
Previously if the SQLiteDatabase.ENABLE_WRITE_AHEAD_LOGGING flag was used when
opening a database, the SQLite synchronous mode would be set to FULL. Update
the default synchronous mode for WAL to be OFF. This will be a huge performance
improvement for databases what use WAL in Robolectric.

For some additional context, the SQLite synchronous mode determines whether or
not SQLite should wait until data is fully written to disk during operations.
Using FULL synchronous mode (the Android default) is important to be resilient
to process crashes. In a Robolectric test environment, however, all SQLite
databases are temporary files, and there is no requirement to be resilient to
crashes.

PiperOrigin-RevId: 393440806