Fix spurious CloseGuard errors in Surface and SurfaceControl
With Robolectric now including stack traces for CloseGuard warnings, there is a
lot of noise associated with CloseGuards for Surface and SurfaceControl objects
initialized in ViewRootImpl, which are used in Activities and Dialogs. The
logic to close the CloseGuards in Surface and SurfaceControl only get called if
the underlying native pointers are non-zero, which means they never get called
in Robolectric. Because of this, the best way to handle this is to explicitly
close any open CloseGuards in the finalize methods of Surface and
SurfaceControl.

PiperOrigin-RevId: 398816996