Use Theme#getResources() to get Resources from SDK 21+
https://github.com/robolectric/robolectric/issues/6757 notifys OpenJDK
18 decides to remove this$ for non-static inner class that doesn't
access outer class. This commit is used to avoid potential problems when
it is shipped to OpenJDK 18, and android-all jars generated by OpenJDK
18.

It will help to remove the usage of "this$0" from SDK 21+. For SDK less
than 21, the andriod-all jars are generated with old JDK some years ago,
we don't need to care about it.

Signed-off-by: utzcoz <utzcoz@outlook.com>