Add a ShadowPackageBackwardCompatibility class to handle when multiple APIs are on the classpath
Specifically, when running a test on API <= P, PackageBackwardCompatibility expects android.content.pm.AndroidTestBaseUpdater to either not be present or extend PackageSharedLibraryUpdater (like it did back then).

But from R onward, android.content.pm.AndroidTestBaseUpdater no longer extends PackageSharedLibraryUpdater. So if we run on API <= P with the R+ AndroidTestBaseUpdater getting loaded by the classloader, PackageBackwardCompatibility#addOptionalUpdater will throw a ClassCastException and fail (since only ClassNotFoundException is being caught). This wasn't found when R support was first added because having the class available was behind a flag (and "not available" was our default), and that flag was removed in S.

PiperOrigin-RevId: 404833512