Always look for tzdata in the corresponding android-all jar instead of the classpath
When running a test on API < S and having API S on the classpath, ShadowMemoryMappedFile would previously pick the tzdata present inside the android-all S jar, but doing so would cause it to error out due to a malformed file (presumably because of a change in format). Having ShadowMemoryMappedFile(S) explicitly use the tzdata from the android-all jar corresponding to the API level of the test fixes this (and is more correct anyway).

PiperOrigin-RevId: 404833509