Synchronize BinderService.getBinder
There was a race that occured if multiple threads simultaneously called
ServiceManager.getService(String) before the service had been initialized. It
was possible that the returned Binder was not fully initialized (i.e.
attachInterface was not called). This results in some internal members of the
Binder being null (specifically mDescriptor). This situation could result in
NPEs when the Binder was used.

PiperOrigin-RevId: 406391901