Add shadows for UwbManager and RangingSession.
The UwbManager is a protocol-agnostic API for managing UWB ranging,
allowing the use of protocol-specific PersistableBundle for configuration.
Then general flow is that for each RangingSession API, a collection of
acceptable parameters are provided by the client, and the framework returns
a collection of selected parameters and a success/failure result via callback.

The general design of this shadow is to intercept the client calls and pass them
to an Adapter registered by a test, such that the test can dictate the result by
calling the client's callback directly, allowing the exercise of varying code paths.
The shadow aims to only couple to public and system APIs and avoid coupling to
internal APIs of the framework.

PiperOrigin-RevId: 408274474