Add the libicu git submodule
Unfortunately, Roboelctric will have to get into the business of performing
custom libicu builds.

As a reminder, libicu is used by Android's SQLite when doing `ORDER BY ...
COLLATE UNICODE` and `ORDER BY ... COLLATE LOCALIZED`.

In general, Robolectric wants to link the static versions of libicu in order to
avoid requiring host machines to have the identical version of libicu that
Robolectric was compiled against. This problem is especially bad on Linux,
where libicu libraries contain their version number as the suffix, almost
guaranteeing that dynamic linking against libicu will not be portable across
different machines.

For Linux, the system versions of libicu do not contain static libraries that
were compiled with `-fPIC`, which means that those static versions of libicu
cannot be linked into a shared library, which is required by the JVM. This
means that Robolectric needs to link against a version of libicu that was
compiled with `-fPIC`.

For Mac, we would like to link against a universal library, mainly because this
will allow us to cross-compile libicu for both Mac M1 and Mac Intel on a single
machine. This will make the CI process much simpler.

Another benefit to having libicu as a submldule is that we can guarantee the
same version of libicu across all host platforms, which could improve
compression ratios of the libicu data files.