Disable the universal library for Mac OS
It is currently not possible to build the universal library on the Mac
GitHub CI runners. Also, building for multiple architectures doubles the
time of compilation for libutc, which can take a very long time on
GitHub CI. It may even be more efficient to have separate mac libraries
to avoid copying an extra 35 MiB when extracting the dylib from
resources for loading.